+++ src/BinaryExtractionTask.spec.ts
@@ -12,8 +12,13 @@ describe("BinaryExtractionTask", () => {
 
   describe("parser", () => {
     const sut = BinaryExtractionTask.for("ThumbnailImage", "", "")
-    it("returns success from expected input", () => {
-      expect(() => sut.parse("    1 output files created")).to.not.throw()
+    it("returns success (undefined, no error) from expected input", () => {
+      expect(sut.parse("    1 output files created")).to.eql(undefined)
+    })
+    it("returns error from expected input", () => {
+      expect(sut.parse("     0 output files created")).to.eql(
+        "0 output files created"
+      )
     })
     it("throws on empty input", () => {
       expect(() => sut.parse("")).to.throw(/Missing expected status message/)
@@ -21,11 +26,9 @@ describe("BinaryExtractionTask", () => {
     it("extracts the expected error message", () => {
       expect(() =>
         sut.parse(
-          [
-            "Error creating /etc/test.jpg",
-            "1 files could not be read",
-            "0 output files created",
-          ].join("\n")
+          ["Error creating /etc/test.jpg", "1 files could not be read"].join(
+            "\n"
+          )
         )
       ).to.throw(/Error creating \/etc\/test.jpg/)
     })
@@ -38,6 +41,18 @@ describe("BinaryExtractionTask", () => {
     await exiftool.extractThumbnail(src, dest)
     // exiftool with_thumb.jpg -b -ThumbnailImage | sha1sum
     expect(await sha1(dest)).to.eql("c7c14706fce4038f6a9da96e213768756a4b2ad2")
+  })
+
+  it("throws for missing src", async function () {
+    this.slow(500)
+    const src = join(testDir, "nonexistant-file.jpg")
+    const dest = await tmpname()
+    try {
+      await exiftool.extractJpgFromRaw(src, dest)
+      fail("expected error to be thrown")
+    } catch (err) {
+      expect(String(err)).to.match(/File not found/i)
+    }
   })
 
   it("throws for missing thumb", async function () {